remove tag schema/controller and old user schema (idk why its still there..)
