+++ mobile/Features/ContentLibrary/Components/Card.jsx
@@ -4,6 +4,7 @@ import {
 import PropTypes from 'prop-types';
 import { useContext } from 'react';
 import axios from 'axios';
+import { useSelector } from 'react-redux';
 import style from './ContentStyle';
 import starImage from '../../../assets/images/star.png';
 import goldStar from '../../../assets/images/goldStar.png';
@@ -16,6 +17,7 @@ export default function Card({
   const {
     deleteFavorite, addFavorite, findFavorite,
   } = useContext(TagContext);
+  const { email, authHeader } = useSelector((state) => state.auth);
 
   /* Grab item fields */
   const { _id, tagName } = item;
@@ -26,13 +28,13 @@ export default function Card({
   /* Adds Tag to Users Favorites List */
   const favoriteTag = async () => {
     addFavorite(_id);
-    await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/tag/favoriteTag`, { tag: { id: _id, tagName }, email: 'hi' });
+    await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/offUser/favoriteTag`, { tag: { id: _id, tagName }, email }, { headers: authHeader });
   };
 
   /* Remove Tag from Users Favorites List */
   const unfavoriteTag = async () => {
     deleteFavorite(_id);
-    await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/tag/unfavoriteTag`, { tag: { id: _id, tagName }, email: 'hi' });
+    await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/offUser/unfavoriteTag`, { tag: { id: _id, tagName }, email }, { headers: authHeader });
   };
 
   /* Handles Favorite Change */
