+++ mobile/Features/ContentLibrary/Screens/Tag.jsx
@@ -5,6 +5,7 @@ import PropTypes from 'prop-types';
 import Swiper from 'react-native-swiper';
 import axios from 'axios';
 import { useContext } from 'react';
+import { useSelector } from 'react-redux';
 import starImage from '../../../assets/images/star.png';
 import leftArrow from '../../../assets/images/left.png';
 import rightArrow from '../../../assets/images/right.png';
@@ -17,6 +18,7 @@ import TagContext from '../Context/TagContext';
 export default function Tag({ navigation, route }) {
   /* index of corresponding Tag */
   const { index, routeName } = route.params;
+  const { email, authHeader } = useSelector((state) => state.auth);
 
   const {
     Tags, deleteFavorite, addFavorite, findFavorite,
@@ -39,13 +41,13 @@ export default function Tag({ navigation, route }) {
   /* Adds Tag to Users Favorites List */
   const favoriteTag = async () => {
     addFavorite(_id);
-    await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/tag/favoriteTag`, { tag: { id: _id, tagName }, email: 'hi' });
+    await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/offUser/favoriteTag`, { tag: { id: _id, tagName }, email }, { headers: authHeader });
   };
 
   /* Remove Tag from Users Favorites List */
   const unfavoriteTag = async () => {
     deleteFavorite(_id);
-    await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/tag/unfavoriteTag`, { tag: { id: _id, tagName }, email: 'hi' });
+    await axios.post(`${process.env.EXPO_PUBLIC_SERVER_URL}/offUser/unfavoriteTag`, { tag: { id: _id, tagName }, email }, { headers: authHeader });
   };
 
   /* Handles Favorite Change */
