+++ server/routes/offUserRoute.js
@@ -35,4 +35,10 @@ offUserRouter.post('/increaseChallengeDay', offUserController.increaseChallengeD
 
 offUserRouter.post('/readSpecifiedFields', offUserController.readSpecifiedFields);
 
+offUserRouter.post('/getFavorites', offUserController.getFavorites);
+
+offUserRouter.post('/favoriteTag', offUserController.favoriteTag);
+
+offUserRouter.post('/unfavoriteTag', offUserController.unfavoriteTag);
+
 module.exports = offUserRouter;
